Scrape west yorkshire locations
Very basic use of the `uk_planning_scraper` gem for demonstration purposes. Can be expanded much much further, but it's a state.
N.B: There are a LOT more comments in this class than I'd be accustomed to making, but this is for educational purposes. These will eventually be moved into the `README.md` in a later commit for proper documentation purposes.